api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pyyaml==6.0.2
requests==2.32.3
redis==5.0.8
pandas==2.2.2
numpy==1.26.4
python-multipart==0.0.12
pydantic==2.9.2
